Derive receive timeout from negotiated session timeout
The receive timeout is first calculated as 2/3 of the session timeout value
that the user wants. However, after connecting to a client, the session timeout
is potentially changed, while the receive timeout stays relative to the inital
value. This commit changes the receive timeout to stay at 2/3 of the negotiated
session timeout after connecting. This behavior is in line with the Java
Zookeeper client.