Add an option to supply a callback to be run for each Event.
Background:

The channel approach causes lost events during period of scheduling contention.
Of particular harm is the loss of HasSession since it's usually the state that
triggers actions.

The callback approch puts the responsiblity on the caller of the library to
decide on the tradeoff between losing events and blocking the main event loop.