add WithLogger connOption
This allows setting the logger as the connection is created, rather
than racing to call SetLogger while the connection's goroutine might
already be logging a successful (or failed) connection.

In addition using WithLogger works around the data race between SetLogger
and the connection's goroutine.

The data race is NOT fixed by this commit. AFAICS doing so is a nontrivial.
Every access to conn.logger would need to be atomic.