WithLogger connOption and Conn.logInfo bool (#170)
* add WithLogger connOption

This allows setting the logger as the connection is created, rather
than racing to call SetLogger while the connection's goroutine might
already be logging a successful (or failed) connection.

In addition using WithLogger works around the data race between SetLogger
and the connection's goroutine.

The data race is NOT fixed by this commit. AFAICS doing so is a nontrivial.
Every access to conn.logger would need to be atomic.

* add Conn.logInfo bool controling informational messages

By default it is enabled for backwards compatability, but
by setting logInfo=false you can have silence on stderr
when no errors are in fact occurring.

This is desireable for commandline tools, where the convention
I prefer is that silence indicates success.